(refactor): collapse RedisBase into RedisTest
